+++ src/buy_function.py
@@ -1,29 +1,16 @@
-import os
 import json
-from typing import Final
 
 from cbpro import AuthenticatedClient
 from aws_lambda_powertools.logging import Logger
 
-from commons import get_balance, InsufficientFundsException
+from src.commons import init_cbpro_client, get_balance, InsufficientFundsException
 
-logging = Logger()
-
-API_SECRET: Final[str] = os.environ['API_SECRET']
-API_KEY: Final[str] = os.environ['API_KEY']
-API_PASS: Final[str] = os.environ['API_PASS']
-URL: Final[str] = os.environ['URL']
-
-client = AuthenticatedClient(
-    key=API_KEY,
-    b64secret=API_SECRET,
-    passphrase=API_PASS,
-    api_url=URL
-)
+client: AuthenticatedClient = init_cbpro_client()
+logger = Logger()
 
 
 def lambda_handler(event, context):
-    logging.info({
+    logger.info({
         'event': event,
         'context': context,
     })
@@ -33,7 +20,7 @@ def lambda_handler(event, context):
         target_currency='BTC'
     )
 
-    logging.info({
+    logger.info({
         'account_id': account_id,
         'balance': balance,
     })
@@ -48,10 +35,10 @@ def lambda_handler(event, context):
             'order_details': order_details,
         }
 
-        logging.info(result)
+        logger.info(result)
         return {
             'statusCode': 200,
-            'body': json.dumps(result, default=str)
+            'body': json.dumps(result)
         }
 
     raise InsufficientFundsException(f'{account_id=}; {balance=}')
